feat: tg-44 #deployment, added promotions

+++ app/src/main/kotlin/com/grappim/taigamobile/createtask/CreateWorkItemData.kt
@@ -2,4 +2,4 @@ package com.grappim.taigamobile.createtask
 
 import com.grappim.taigamobile.core.domain.CommonTaskType
 
-data class CreateWorkItemData(val id: Long, val type: CommonTaskType, val ref: Int)
+data class CreateWorkItemData(val id: Long, val type: CommonTaskType, val ref: Long)
