+++ feature/epics/dto/src/main/java/com/grappim/taigamobile/feature/epics/dto/EpicShortInfoDTO.kt
@@ -4,4 +4,4 @@ import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class EpicShortInfoDTO(val id: Long, @Json(name = "subject") val title: String, val ref: Int, val color: String)
+data class EpicShortInfoDTO(val id: Long, @Json(name = "subject") val title: String, val ref: Long, val color: String)
