+++ feature/userstories/domain/src/main/java/com/grappim/taigamobile/feature/userstories/domain/UserStoryEpic.kt
@@ -1,3 +1,3 @@
 package com.grappim.taigamobile.feature.userstories.domain
 
-data class UserStoryEpic(val id: Long, val title: String, val ref: Int, val color: String)
+data class UserStoryEpic(val id: Long, val title: String, val ref: Long, val color: String)
