+++ feature/workitem/domain/src/main/java/com/grappim/taigamobile/feature/workitem/domain/UserStoryShortInfo.kt
@@ -2,4 +2,4 @@ package com.grappim.taigamobile.feature.workitem.domain
 
 import kotlinx.collections.immutable.ImmutableList
 
-data class UserStoryShortInfo(val id: Long, val ref: Int, val title: String, val epics: ImmutableList<EpicShortInfo>)
+data class UserStoryShortInfo(val id: Long, val ref: Long, val title: String, val epics: ImmutableList<EpicShortInfo>)
